moved create account sheet to the styles folder
