--- src/components/AuthenticationMainText.css
@@ -1,14 +1,0 @@
-.mainTextContainer {
-    display: flex;
-    flex-direction: column;
-}
-
-.mainTextHeader {
-    font-size: 3.3em;
-    color:#00151A;
-}
-
-.mainTextBody {
-    font-size: 1.5em ;
-    color: #55666A;
-}
